Theme almost works

+++ Models/Debtor.cs
@@ -7,7 +7,7 @@ public class Debtor
     [PrimaryKey, AutoIncrement]
     public int Id { get; set; }
     public string Name { get; set; }
-    public decimal AmountOwed { get; set; }
+    public double AmountOwed { get; set; }
 
     
 }
